requirements.txt
fastapi[standard]
uvicorn
streamlit
networkx
chromadb
sentence-transformers
pandas
pyarrow
python-dotenv
matplotlib
pytest
requests
langchain
langgraph
langchain-community
langchain-chroma
langchain-text-splitters
langchain-huggingface
langchain-openai
langchain-experimental


